+++ scripts/ci/index_ref_doc.py
@@ -14,7 +14,8 @@ import traceback
 import unittest
 import shutil
 from subprocess import check_call, CalledProcessError
-from pkg_resources import parse_version, get_distribution
+from packaging.version import Version
+from importlib.metadata import version as get_version
 
 from six import with_metaclass
 
@@ -31,18 +32,18 @@ if not os.path.isdir(REF_DOC_OUT_DIR):
 
 ALL_TESTS = []
 
-CLI_VERSION = get_distribution('azure-cli').version
+CLI_VERSION = get_version('azure-cli')
 
 for extension_name, exts in get_index_data()['extensions'].items():
-    parsed_cli_version = parse_version(CLI_VERSION)
+    parsed_cli_version = Version(CLI_VERSION)
     filtered_exts = []
     for ext in exts:
-        if parsed_cli_version <= parse_version(ext['metadata'].get('azext.maxCliCoreVersion', CLI_VERSION)):
+        if parsed_cli_version <= Version(ext['metadata'].get('azext.maxCliCoreVersion', CLI_VERSION)):
             filtered_exts.append(ext)
     if not filtered_exts:
         continue
 
-    candidates_sorted = sorted(filtered_exts, key=lambda c: parse_version(c['metadata']['version']), reverse=True)
+    candidates_sorted = sorted(filtered_exts, key=lambda c: Version(c['metadata']['version']), reverse=True)
     chosen = candidates_sorted[0]
     ALL_TESTS.append((extension_name, chosen['downloadUrl'], chosen['filename']))