+++ src/confcom/HISTORY.rst
@@ -3,6 +3,14 @@
 Release History
 ===============
 
+1.7.1
+++++++
+* Replace deprecated pkg_resources with packaging for Python 3.13 compatibility
+
+1.7.0
+++++++
+* Add containers from_vn2 command to generate container definitions from a VN2 template.
+
 1.6.0
 ++++++
 * Added confcom containers from_image command to generate container definitions from an image reference